finished

+++ average.py
@@ -3,9 +3,9 @@
 #Date: 6/30/2025
 #Description: Takes 5 user number inputs and averages them
 print ("Please enter 5 numbers for me to average")
-a = int(input())
-b = int(input())
-c = int(input())
-d = int(input())
-e = int(input())
+a = float(input())
+b = float(input())
+c = float(input())
+d = float(input())
+e = float(input())
 print((a + b + c + d + e)/5)
